+++ pronamic-pay-woocommerce.php
@@ -4,8 +4,9 @@
  * Plugin URI: https://www.pronamic.eu/plugins/pronamic-pay-woocommerce/
  * Description: Extend the Pronamic Pay plugin with WooCommerce support to receive payments through a variety of payment providers.
  *
- * Version: 4.3.2
- * Requires at least: 4.7
+ * Version: 4.5.5
+ * Requires at least: 5.9
+ * Requires PHP: 7.4
  *
  * Author: Pronamic
  * Author URI: https://www.pronamic.eu/
@@ -15,12 +16,13 @@
  *
  * License: GPL-3.0-or-later
  *
+ * Requires Plugins: pronamic-ideal, woocommerce
  * Depends: wp-pay/core
  *
  * GitHub URI: https://github.com/pronamic/wp-pronamic-pay-woocommerce
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2022 Pronamic
+ * @copyright 2005-2023 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\WooCommerce
  */