+++ src/Gateway.php
@@ -3,7 +3,7 @@
  * Gateway
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2022 Pronamic
+ * @copyright 2005-2023 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\WooCommerce
  */
@@ -13,16 +13,17 @@ namespace Pronamic\WordPress\Pay\Extensions\WooCommerce;
 use Pronamic\WordPress\DateTime\DateTime;
 use Pronamic\WordPress\Money\Money;
 use Pronamic\WordPress\Money\TaxedMoney;
+use Pronamic\WordPress\Number\Number;
 use Pronamic\WordPress\Pay\Address;
 use Pronamic\WordPress\Pay\Core\Field;
 use Pronamic\WordPress\Pay\Customer;
 use Pronamic\WordPress\Pay\ContactName;
 use Pronamic\WordPress\Pay\Core\PaymentMethods;
-use Pronamic\WordPress\Pay\Core\Util;
 use Pronamic\WordPress\Pay\Payments\Payment;
 use Pronamic\WordPress\Pay\Payments\PaymentLines;
 use Pronamic\WordPress\Pay\Payments\PaymentLineType;
 use Pronamic\WordPress\Pay\Payments\PaymentStatus;
+use Pronamic\WordPress\Pay\Refunds\Refund;
 use Pronamic\WordPress\Pay\Plugin;
 use Pronamic\WordPress\Pay\Region;
 use Pronamic\WordPress\Pay\Subscriptions\Subscription;
@@ -32,7 +33,7 @@ use WC_Payment_Gateway;
 /**
  * Title: WooCommerce iDEAL gateway
  * Description:
- * Copyright: 2005-2022 Pronamic
+ * Copyright: 2005-2023 Pronamic
  * Company: Pronamic
  *
  * @link https://github.com/woocommerce/woocommerce/blob/3.5.3/includes/abstracts/abstract-wc-payment-gateway.php
@@ -636,6 +637,8 @@ class Gateway extends WC_Payment_Gateway {
 
 		$payment = new Payment();
 
+		$payment->set_meta( 'woocommerce_order_id', $order->get_id() );
+
 		/*
 		 * An '#' character can result in the following iDEAL error:
 		 * code             = SO1000
@@ -736,6 +739,8 @@ class Gateway extends WC_Payment_Gateway {
 		 */
 		$items = $order->get_items( [ 'line_item', 'fee', 'shipping' ] );
 
+		$tax_percentages = [ 0 ];
+
 		$payment->lines = new PaymentLines();
 
 		foreach ( $items as $item_id => $item ) {
@@ -750,17 +755,23 @@ class Gateway extends WC_Payment_Gateway {
 				$quantity = 1;
 			}
 
+			// Tax.
+			$tax_rate_id = WooCommerce::get_order_item_tax_rate_id( $item );
+
+			$percent = is_null( $tax_rate_id ) ? null : \WC_Tax::get_rate_percent_value( $tax_rate_id );
+
 			// Set line properties.
-			$line->set_id( $item_id );
+			$line->set_id( (string) $item_id );
 			$line->set_sku( WooCommerce::get_order_item_sku( $item ) );
 			$line->set_type( (string) $type );
 			$line->set_name( $item['name'] );
 			$line->set_quantity( $quantity );
-			$line->set_unit_price( new TaxedMoney( $order->get_item_total( $item, true ), WooCommerce::get_currency(), $order->get_item_tax( $item ) ) );
-			$line->set_total_amount( new TaxedMoney( $order->get_line_total( $item, true ), WooCommerce::get_currency(), $order->get_line_tax( $item ) ) );
+			$line->set_unit_price( new TaxedMoney( $order->get_item_total( $item, true ), WooCommerce::get_currency(), $order->get_item_tax( $item ), $percent ) );
+			$line->set_total_amount( new TaxedMoney( $order->get_line_total( $item, true ), WooCommerce::get_currency(), $order->get_line_tax( $item ), $percent ) );
 			$line->set_product_url( WooCommerce::get_order_item_url( $item ) );
 			$line->set_image_url( WooCommerce::get_order_item_image( $item ) );
 			$line->set_product_category( WooCommerce::get_order_item_category( $item ) );
+			$line->set_meta( 'woocommerce_order_item_id', $item_id );
 		}
 
 		return $payment;
@@ -801,7 +812,7 @@ class Gateway extends WC_Payment_Gateway {
 
 	/**
 	 * Connection subscription payment renewal.
-	 * 
+	 *
 	 * @param Payment  $payment Payment.
 	 * @param WC_Order $order   WooCommerce order.
 	 * @return void
@@ -889,7 +900,7 @@ class Gateway extends WC_Payment_Gateway {
 
 	/**
 	 * Maybe add subscriptions support.
-	 * 
+	 *
 	 * @return void
 	 */
 	public function maybe_add_subscriptions_support() {
@@ -931,21 +942,80 @@ class Gateway extends WC_Payment_Gateway {
 
 		$amount = new Money( $amount, $order->get_currency( 'raw' ) );
 
-		try {
-			$refund_reference = Plugin::create_refund( $order->get_transaction_id(), $gateway, $amount, $reason );
+		$payment_id = $order->get_meta( '_pronamic_payment_id' );
 
-			if ( null !== $refund_reference ) {
-				$note = \sprintf(
-					/* translators: 1: formatted refund amount, 2: refund gateway reference */
-					\__( 'Created refund of %1$s with gateway reference `%2$s`.', 'pronamic_ideal' ),
-					\esc_html( $amount->format_i18n() ),
-					\esc_html( $refund_reference )
-				);
+		$payment = \get_pronamic_payment( $payment_id );
 
-				$order->add_order_note( $note );
+		if ( null === $payment ) {
+			return new \WP_Error(
+				'pronamic-pay-woocommerce-refund-payment',
+				\__( 'Cannot process refund because payment could not be found.', 'pronamic_ideal' )
+			);
+		}
+
+		$payment_lines = $payment->get_lines();
+
+		$refund = new Refund( $payment, $amount );
+
+		$refund->created_by = \wp_get_current_user();
+
+		$refund->set_description( $reason );
+
+		$refunds = $order->get_refunds();
+
+		$refund_order = reset( $refunds );
+
+		if ( false !== $refund_order ) {
+			$items = $refund_order->get_items( [ 'line_item', 'fee', 'shipping' ] );
+
+			foreach ( $items as $item_id => $item ) {
+				$line = $refund->lines->new_line();
+
+				$type = OrderItemType::transform( $item );
+
+				// Quantity.
+				$quantity = wc_stock_amount( $item['qty'] );
+
+				if ( PaymentLineType::SHIPPING === $type ) {
+					$quantity = -1;
+				}
+
+				// Tax.
+				$tax_rate_id = WooCommerce::get_order_item_tax_rate_id( $item );
+
+				$percent = is_null( $tax_rate_id ) ? null : \WC_Tax::get_rate_percent_value( $tax_rate_id );
+
+				// Set line properties.
+				$line->set_id( $item_id );
+				$line->set_quantity( Number::from_mixed( $quantity )->negative() );
+				$line->set_total_amount( new TaxedMoney( -1 * $refund_order->get_line_total( $item, true ), WooCommerce::get_currency(), -1 * $refund_order->get_line_tax( $item ), $percent ) );
+				$line->set_meta( 'woocommerce_refunded_item_id', $item->get_meta( '_refunded_item_id' ) );
+
+				if ( null !== $payment_lines ) {
+					$payment_line = $payment_lines->first( $item->get_meta( '_refunded_item_id' ) );
+
+					if ( null !== $payment_line ) {
+						$line->meta = $payment_line->meta;
+						
+						$line->set_payment_line( $payment_line );
+					}
+				}
 			}
 
-			$order->update_meta_data( '_pronamic_amount_refunded', (string) $amount->get_value() );
+			$refund->meta['woocommerce_order_id'] = $refund_order->get_id();
+		}
+
+		try {
+			Plugin::create_refund( $refund );
+
+			$note = \sprintf(
+				/* translators: 1: formatted refund amount, 2: refund gateway reference */
+				\__( 'Created refund of %1$s with reference `%2$s`.', 'pronamic_ideal' ),
+				\esc_html( $amount->format_i18n() ),
+				\esc_html( $refund->psp_id )
+			);
+
+			$order->add_order_note( $note );
 		} catch ( \Exception $e ) {
 			return new \WP_Error(
 				'pronamic-pay-woocommerce-refund',
@@ -1031,11 +1101,15 @@ class Gateway extends WC_Payment_Gateway {
 			foreach ( $fields as $field ) {
 				echo '<p class="form-row form-row-wide">';
 
-				\printf(
-					'<label for="%s">%s</label> ',
-					\esc_attr( $field->get_id() ),
-					\esc_html( $field->get_label() )
-				);
+				$label = $field->get_label();
+
+				if ( ! empty( $label ) ) {
+					\printf(
+						'<label for="%s">%s</label> ',
+						\esc_attr( $field->get_id() ),
+						\esc_html( $label )
+					);
+				}
 
 				try {
 					$field->output();
@@ -1074,11 +1148,11 @@ class Gateway extends WC_Payment_Gateway {
 		foreach ( $input_ids as $input_id ) {
 			$input_name = sprintf( '%s_%s', $this->id, $input_id );
 
-			if ( ! filter_has_var( INPUT_POST, $input_name ) ) {
+			if ( ! array_key_exists( $input_name, $data ) ) {
 				continue;
 			}
 
-			$input_value = filter_input( INPUT_POST, $input_name, FILTER_SANITIZE_STRING );
+			$input_value = $data[ $input_name ];
 
 			// Add error for empty input value.
 			if ( empty( $input_value ) ) {